+++ cmd/internal/provider/default_systems.go
@@ -5,11 +5,13 @@ import (
 
 	"github.com/google/trillian/quota"
 	"github.com/google/trillian/storage"
+	"github.com/google/trillian/util/election2"
 )
 
 var (
-	DefaultQuotaSystem   string
-	DefaultStorageSystem string
+	DefaultQuotaSystem    string
+	DefaultStorageSystem  string
+	DefaultElectionSystem string
 )
 
 func init() {
@@ -27,4 +29,12 @@ func init() {
 		defaultProvider = providers[0]
 	}
 	DefaultQuotaSystem = defaultProvider
+
+	defaultProvider = "etcd"
+	providers = election2.Providers()
+	if len(providers) > 0 && !slices.Contains(providers, defaultProvider) {
+		slices.Sort(providers)
+		defaultProvider = providers[0]
+	}
+	DefaultElectionSystem = defaultProvider
 }